plugins/requirements.txt
urllib3==1.22
python-dateutil==2.6.1
Pillow==4.0.0
dicttoxml==1.7.4
GoogleAppEngineCloudStorageClient==1.9.22.1